pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.95.0",
    "atlassian-python-api>=4.0.7",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.136.0",
    "mcp>=1.26.0",
    "openai>=2.32.0",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]